requirements.txt
# This is a PIP requirements file.
# To setup a dev environment:
#
#   easy_install virtualenv
#   virtualenv --distribute --no-site-packages --prompt "(graphite venv) " .venv
#   source .venv/bin/activate
#
#   brew install cairo && brew link cairo # on OSX
#   pip install -r requirements.txt
#
#   mkdir -p .venv/storage .venv/conf
#   touch .venv/conf/carbon.conf .venv/conf/storage-schemas.conf
#   cp conf/graphTemplates.conf.example .venv/conf/graphTemplates.conf
#   export GRAPHITE_ROOT=`pwd` GRAPHITE_STORAGE_DIR=`pwd`/.venv/storage/ GRAPHITE_CONF_DIR=`pwd`/.venv/conf/
#
#   carbon/bin/carbon-cache.py start
#   carbon/bin/carbon-cache.py status
#
#   mkdir -p .venv/storage/log/webapp
#   .venv/bin/django-admin.py syncdb --settings=graphite.settings --pythonpath=webapp
#   bin/run-graphite-devel-server.py ./
#   # or
#   # cd webapp/graphite && $GRAPHITE_ROOT/.venv/bin/gunicorn_django -b 127.0.0.1:8080
#
# To clean up:
#
#   carbon/bin/carbon-cache.py stop
#   unset GRAPHITE_ROOT GRAPHITE_STORAGE_DIR GRAPHITE_CONF_DIR
#   deactivate
#

Django==1.3
Twisted==11.0.0
python-memcached==1.47
txAMQP==0.4
simplejson==2.1.6
django-tagging==0.3.1
gunicorn
pytz
http://cairographics.org/releases/py2cairo-1.8.10.tar.gz
git+git://github.com/graphite-project/whisper.git#egg=whisper
